Add relationship model and skeleton view
The relationship model consists of 2 authors, author1 and author2,
and a relationship status of true or false.

Having a row of:

    author1, author2, false

means author1 sent a friend request to author2 but author2 has not
accepted it yet.

    author1, author2, true

means author1 sent a friend request to author2 and author2 has
accepted it.

If a there is no row with author1, author2, it means neither author
has sent a friend request to each other.