Minor HTML cleanup
I have my reservations about base.css.  I dislike the float: left on all divs, as it is causing a lot of needing clear-lefts and spans just to counteract it all.  But for now, I have tried to minimize any use of classes so there is less pain in refactoring later.