Initial commit of image hosting support (Closed #60)
Sorry, but you might need to nuke your DBs for this one.

Dogenode now supports image hosting.  Still need to work on allowing
users to upload files, but if you manually upload images via the admin
panel, you can verify that it works.

If a user cannot view an image, not_allowed.png is shown instead.  If
a file is not found, not_found.png is shown instead.