GitHub stream tests implemented, fixed a really glaringly bad issue
Man, how stupid of me.  I pulled all posts for a user before querying GitHub for any
events, so that's why my tests were failing on me.  It seems that when Django is not
running in test mode, the "Get allowed posts" query is delayed until the very last minute,
i.e. when we start looping through its QuerySet.  However, in test mode, the queries seem
to be executed immediately, which goes to show that it reveals really bad code really quickly.

Anyway hello #76