Implemented new and more reliable way to share images
Fixed #84.  Hallelujah!

Now, I expect the user to upload the images prior to sharing them.  Initially, all uploaded
and unassociated images are private.  Once they are linked to a post, then the image adopts
all visibility and visibility exceptions of that post.  One downside to this implementation
is that for a user to share the same image in multiple posts, he/she will have to re-upload
the image.  However, this is a small price to pay for ensuring that images always work,
regardless of where Dogenode is hosted.

Basically, POST is still handled way, way better in browsers and servers than PUT is when
it comes to handling binary data.

There is one unit test that needs to be fixed, but I will do that @sgil88 @brontel!